42 us different price according to user type

+++ src/main/java/sophiatech/PaymentService.java
@@ -1,7 +1,7 @@
 package sophiatech;
 
 public class PaymentService {
-    public boolean pay(int total) {
+    public boolean pay(double total) {
         //fake payment service for now, will accept everything
         return true;
     }
